requirements.txt
playwright==1.44.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
alembic==1.13.1
fastapi==0.111.0
uvicorn==0.29.0
apscheduler==3.10.4
sentence-transformers==2.7.0
spacy==3.7.4
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1
loguru==0.7.2
aiofiles==23.2.1
